blog/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
bleach==3.3.1
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-haystack==3.0
django-imagekit==4.0.2
django-mysql==3.12.0
django-redis==5.0.0
django-reversion==4.0.0
django-tctip==1.4.2
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
html5lib==1.1
idna==3.2
kombu==5.1.0
Markdown==3.3.4
oauthlib==3.1.1
packaging==21.0
pilkit==2.0
Pillow==8.3.1
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
